requirements.txt
Django==2.0.3
djangorestframework
